third_party/WebKit/LayoutTests/fast/events/platform-wheelevent-paging-x-in-non-scrolling-div-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS event.wheelDeltaY is 0
PASS event.wheelDeltaX is window.givenScrollLeft*-3
PASS event.wheelDelta is window.givenScrollLeft*-3
PASS div.scrollTop is window.expectedScrollTop
PASS div.scrollLeft is window.expectedScrollLeft
